move Ops into their own package and import cleanup
